requirements.txt
requests==2.32.4
pycryptodome==3.23.0
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1